+++ src/shapes/IText/ITextClickBehavior.ts
@@ -239,7 +239,7 @@ export abstract class ITextClickBehavior<
    * @return {Number} Index of a character
    */
   getSelectionStartFromPointer(e: TPointerEvent): number {
-    const mouseOffset = this.canvas!.getPointer(e)
+    const mouseOffset = this.canvas!.getScenePoint(e)
       .transform(invertTransform(this.calcTransformMatrix()))
       .add(new Point(-this._getLeftOffset(), -this._getTopOffset()));
     let height = 0,