+++ src/shapes/IText/DraggableTextDelegate.ts
@@ -124,7 +124,7 @@ export class DraggableTextDelegate {
     const diff = pointer.subtract(pos);
     const enableRetinaScaling = canvas._isRetinaScaling();
     const retinaScaling = target.getCanvasRetinaScaling();
-    const bbox = target.getBoundingRect(true);
+    const bbox = target.getBoundingRect();
     const correction = pos.subtract(new Point(bbox.left, bbox.top));
     const vpt = canvas.viewportTransform;
     const offset = correction.add(diff).transform(vpt, true);