+++ src/shapes/IText/DraggableTextDelegate.ts
@@ -123,7 +123,7 @@ export class DraggableTextDelegate {
     const pointer = canvas.getScenePoint(e);
     const diff = pointer.subtract(pos);
     const retinaScaling = target.getCanvasRetinaScaling();
-    const bbox = target.getBoundingRect(true);
+    const bbox = target.getBoundingRect();
     const correction = pos.subtract(new Point(bbox.left, bbox.top));
     const vpt = canvas.viewportTransform;
     const offset = correction.add(diff).transform(vpt, true);