+++ src/shapes/Object/Object.ts
@@ -1375,8 +1375,8 @@ export class FabricObject<
     }
 
     const el = createCanvasElement(),
-      // skip canvas zoom and calculate with setCoords now.
-      boundingRect = this.getBoundingRect(true, true),
+      // calculate with setCoords now.
+      boundingRect = this.getBoundingRect(true),
       shadow = this.shadow,
       shadowOffset = new Point();