+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [next]
 
+- docs(): DRAFT: Remove inheritance from docs [#9486](https://github.com/fabricjs/fabric.js/pull/9486)
 - chore(): expose `sendVectorToPlane` [#9479](https://github.com/fabricjs/fabric.js/pull/9479)
 - BREAKING: remove absolute true/false from the api. [#9395](https://github.com/fabricjs/fabric.js/pull/9395)
 - refactor(Canvas): BREAKING deprecate `getPointer`, add new getScenePoint and getViewportPoint methods, removed `restorePointerVpt`, extended mouse events data [#9175](https://github.com/fabricjs/fabric.js/pull/9175)